# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Asset Management API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/lokasi-assets": {
            "get": {
                "tags": [
                    "LokasiAssets"
                ],
                "summary": "List all asset locations",
                "operationId": "1ec5703fd999778b4942ec43c4a402e6",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LokasiAsset"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "LokasiAssets"
                ],
                "summary": "Create a new asset location",
                "operationId": "f4c4b2baac548a7b77e315725ec2cdb3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kode_lokasi",
                                    "nama_lokasi",
                                    "pic_id"
                                ],
                                "properties": {
                                    "kode_lokasi": {
                                        "type": "string",
                                        "example": "LOC002"
                                    },
                                    "nama_lokasi": {
                                        "type": "string",
                                        "example": "Gudang Sparepart"
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "example": "Penyimpanan sparepart"
                                    },
                                    "pic_id": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "asset_count": {
                                        "type": "integer",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lokasi Asset created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LokasiAsset"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lokasi-assets/{id}": {
            "get": {
                "tags": [
                    "LokasiAssets"
                ],
                "summary": "Get detail of an asset location",
                "operationId": "4b98dcb6fa261d1fecdcf4da26908162",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LokasiAsset"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lokasi not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "LokasiAssets"
                ],
                "summary": "Update asset location",
                "operationId": "60e932e80aac87b673bd1c738e6348db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LokasiAsset"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lokasi Asset updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LokasiAsset"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lokasi not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "LokasiAssets"
                ],
                "summary": "Delete asset location",
                "operationId": "bdb31b203ef397a7e2c8a413e0d765e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lokasi Asset deleted"
                    },
                    "404": {
                        "description": "Lokasi not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LokasiAsset": {
                "title": "Lokasi Asset",
                "required": [
                    "kode_lokasi",
                    "nama_lokasi",
                    "pic_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "kode_lokasi": {
                        "type": "string",
                        "example": "LOC001"
                    },
                    "nama_lokasi": {
                        "type": "string",
                        "example": "Gudang Utama"
                    },
                    "keterangan": {
                        "type": "string",
                        "example": "Lokasi untuk penyimpanan utama"
                    },
                    "pic_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "asset_count": {
                        "type": "integer",
                        "example": 120
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}